fix url issue with functions containing underscores
Function names containing underscores were failing the check against the OutputKey
in the stack outputs. For instance, a serverless function named "hello" was properly converted
to "Hello" and matched to "HelloLambdaFunctionQualifiedArn"; however a function named
"hello_world" would be converted to "Hello_world", but the actual arn in the
deployed function (on AWS at least; I did not test other providers) is
"HelloUnderscoreworldLambdaFunctionQualifiedArn"; this led to url's for this function not
appearing in the manifest.